Add test cases to cover the changes
